requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==10.2.0
pygments==2.17.2
requests==2.31.0
rich==13.7.1
urllib3==2.2.1
uuid==1.30
webptools==0.0.9
